+++ store/providers/merkletree/types.go
@@ -12,7 +12,7 @@ type MerkleTree interface {
 	RawAdd(data []byte) (index int, hash string)
 	IntermediaryHashesByIndex(index int) (intermediaryHashes []string, err error)
 	ValidateExistence(original []byte, index int, intermediaryHashes []string) (bool, error)
-	HashAt(index int) (string, error)
+	HashAt(index uint64) (string, error)
 	Root() (*string, error)
 	Length() int
 }
